backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
xrpl-py==2.6.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1
httpx==0.24.1